handle all vars in CTD_CAR2023011_001_496780_DN.ODF.nc
I've checked roughly as I did the work, with
    make | grep ITEM
where ITEM is the file code for something.  The connection of
flags to variables is really quite problematic, though, because
the names do not always match up.  That will need to be documented.
Of course, experts might be working with original names anyway,
so I do not judge this as a blocking problem.